+++ src/calculated/knight.rs
@@ -3,14 +3,14 @@ use std::sync::LazyLock;
 use crate::bitboards::filter;
 use crate::constants::*;
 
-pub fn generate_knight_moves(square: usize) -> u64 {
-    KNIGHT_MOVES[square]
+pub fn generate_knight_moves(square: u8) -> u64 {
+    KNIGHT_MOVES[square as usize]
 }
 
 static KNIGHT_MOVES: LazyLock<[u64; 64]> = LazyLock::new(|| {
     let mut knight_moves = [0; 64];
 
-    for (square, mv) in knight_moves.iter_mut().enumerate().take(H8 + 1) {
+    for (square, mv) in knight_moves.iter_mut().enumerate().take(H8 as usize + 1) {
         let bitboard = 1 << square;
         let mut attacks = 0;
 
@@ -35,33 +35,33 @@ mod tests {
 
     #[test]
     fn full_move_set() {
-        assert_eq!(KNIGHT_MOVES[D3], 0x0014_2200_2214);
-        assert_eq!(KNIGHT_MOVES[E4], 0x2844_0044_2800);
-        assert_eq!(KNIGHT_MOVES[F5], 0x0050_8800_8850_0000);
-        assert_eq!(KNIGHT_MOVES[C6], 0x0a11_0011_0a00_0000);
+        assert_eq!(KNIGHT_MOVES[D3 as usize], 0x0014_2200_2214);
+        assert_eq!(KNIGHT_MOVES[E4 as usize], 0x2844_0044_2800);
+        assert_eq!(KNIGHT_MOVES[F5 as usize], 0x0050_8800_8850_0000);
+        assert_eq!(KNIGHT_MOVES[C6 as usize], 0x0a11_0011_0a00_0000);
     }
 
     #[test]
     fn vertical_edges() {
-        assert_eq!(KNIGHT_MOVES[A4], 0x0204_0004_0200);
-        assert_eq!(KNIGHT_MOVES[B4], 0x0508_0008_0500);
-        assert_eq!(KNIGHT_MOVES[G4], 0xa010_0010_a000);
-        assert_eq!(KNIGHT_MOVES[H4], 0x4020_0020_4000);
+        assert_eq!(KNIGHT_MOVES[A4 as usize], 0x0204_0004_0200);
+        assert_eq!(KNIGHT_MOVES[B4 as usize], 0x0508_0008_0500);
+        assert_eq!(KNIGHT_MOVES[G4 as usize], 0xa010_0010_a000);
+        assert_eq!(KNIGHT_MOVES[H4 as usize], 0x4020_0020_4000);
     }
 
     #[test]
     fn horizontal_edges() {
-        assert_eq!(KNIGHT_MOVES[E1], 0x0028_4400);
-        assert_eq!(KNIGHT_MOVES[E2], 0x2844_0044);
-        assert_eq!(KNIGHT_MOVES[E7], 0x4400_4428_0000_0000);
-        assert_eq!(KNIGHT_MOVES[E8], 0x0044_2800_0000_0000);
+        assert_eq!(KNIGHT_MOVES[E1 as usize], 0x0028_4400);
+        assert_eq!(KNIGHT_MOVES[E2 as usize], 0x2844_0044);
+        assert_eq!(KNIGHT_MOVES[E7 as usize], 0x4400_4428_0000_0000);
+        assert_eq!(KNIGHT_MOVES[E8 as usize], 0x0044_2800_0000_0000);
     }
 
     #[test]
     fn corners() {
-        assert_eq!(KNIGHT_MOVES[A1], 0x20400);
-        assert_eq!(KNIGHT_MOVES[H1], 0x0040_2000);
-        assert_eq!(KNIGHT_MOVES[A8], 0x0004_0200_0000_0000);
-        assert_eq!(KNIGHT_MOVES[H8], 0x0020_4000_0000_0000);
+        assert_eq!(KNIGHT_MOVES[A1 as usize], 0x20400);
+        assert_eq!(KNIGHT_MOVES[H1 as usize], 0x0040_2000);
+        assert_eq!(KNIGHT_MOVES[A8 as usize], 0x0004_0200_0000_0000);
+        assert_eq!(KNIGHT_MOVES[H8 as usize], 0x0020_4000_0000_0000);
     }
 }
